requirements.txt
flask
torch
transformers
pillow
requests